Modularize core logic and persistence
