+++ SDFGridHelpers.js
@@ -3,3 +3,6 @@ export * from './SDFGridConstants.js';
 export * from './SDFGridUtil.js';
 export * from './SDFGridStorage.js';
 export * from './SDFGridNucleus.js';
+export * from './SDFGridPersistence.js';
+export * from './SDFGridLogic.js';
+export * from './SDFGridShape.js';
